+++ Sources/RepoBar/Support/AppActions.swift
@@ -2,11 +2,6 @@ import AppKit
 
 @MainActor
 enum AppActions {
-    static func openSettings() {
-        NSApp.activate(ignoringOtherApps: true)
-        NotificationCenter.default.post(name: .repoBarOpenSettings, object: nil)
-    }
-
     static func openAbout() {
         NSApp.activate(ignoringOtherApps: true)
         NSApp.orderFrontStandardAboutPanel(nil)