+++ Sources/RepoBar/Support/MenuNotifications.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-extension Notification.Name {
-    static let menuFiltersDidChange = Notification.Name("RepoBar.menuFiltersDidChange")
-    static let repoBarOpenSettings = Notification.Name("RepoBar.openSettings")
-}